refactor: rename components and features
