+++ frontend-nextjs/src/features/dashboard/ai-exercise/useExerciseGenerator.ts
